Move notifications code out of BaseActivity to HomeActivity
