+++ app/src/main/java/com/gh4a/activities/home/HomeActivity.java
@@ -3,13 +3,17 @@ package com.gh4a.activities.home;
 import android.app.Dialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.content.ContextCompat;
 import android.support.v4.content.Loader;
+import android.support.v4.graphics.drawable.DrawableCompat;
+import android.support.v4.view.MenuItemCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AlertDialog;
 import android.text.TextUtils;
@@ -26,6 +30,7 @@ import com.gh4a.R;
 import com.gh4a.activities.Github4AndroidActivity;
 import com.gh4a.activities.SettingsActivity;
 import com.gh4a.activities.UserActivity;
+import com.gh4a.fragment.NotificationListFragment;
 import com.gh4a.fragment.RepositoryListContainerFragment;
 import com.gh4a.fragment.SettingsFragment;
 import com.gh4a.loader.HasNotificationsLoader;
@@ -35,11 +40,13 @@ import com.gh4a.loader.UserLoader;
 import com.gh4a.utils.ApiHelpers;
 import com.gh4a.utils.AvatarHandler;
 import com.gh4a.utils.IntentUtils;
+import com.gh4a.utils.UiUtils;
 
 import org.eclipse.egit.github.core.User;
 
 public class HomeActivity extends BasePagerActivity implements
-        View.OnClickListener, RepositoryListContainerFragment.Callback {
+        View.OnClickListener, RepositoryListContainerFragment.Callback,
+        NotificationListFragment.ParentCallback {
     private static final int REQUEST_SETTINGS = 10000;
 
     private FragmentFactory mFactory;
@@ -51,6 +58,10 @@ public class HomeActivity extends BasePagerActivity implements
     private int mSelectedFactoryId;
     private boolean mDrawerInAccountMode;
     private Menu mLeftDrawerMenu;
+    private ImageView mNotificationsIndicator;
+    private MenuItem mNotificationsMenuItem;
+    private Drawable mTintedCircleIcon;
+    private Drawable mCircleIcon;
 
     private static final String STATE_KEY_FACTORY_ITEM = "factoryItem";
 
@@ -105,6 +116,8 @@ public class HomeActivity extends BasePagerActivity implements
         }
         mFactory = getFactoryForItem(mSelectedFactoryId);
 
+        setupIndicatorIcon();
+
         super.onCreate(savedInstanceState);
 
         ActionBar actionBar = getSupportActionBar();
@@ -115,6 +128,32 @@ public class HomeActivity extends BasePagerActivity implements
 
         getSupportLoaderManager().initLoader(0, null, mUserCallback);
         getSupportLoaderManager().initLoader(1, null, mHasNotificationsCallback);
+    }
+
+    private void setupIndicatorIcon() {
+        int circleIconRes = UiUtils.resolveDrawable(this, R.attr.circleIcon);
+
+        mCircleIcon = DrawableCompat.wrap(ContextCompat.getDrawable(this, circleIconRes).mutate());
+        DrawableCompat.setTint(mCircleIcon, UiUtils.resolveColor(this, android.R.attr.textColorPrimary));
+
+        mTintedCircleIcon = DrawableCompat.wrap(ContextCompat.getDrawable(this, circleIconRes).mutate());
+        DrawableCompat.setTint(mTintedCircleIcon, UiUtils.resolveColor(this, R.attr.colorAccent));
+    }
+
+    private void updateNotificationIndicator(int checkedItemId) {
+        if (mNotificationsIndicator != null) {
+            mNotificationsIndicator.setImageDrawable(
+                    checkedItemId == R.id.notifications ? mTintedCircleIcon : mCircleIcon);
+        }
+    }
+
+    public void setNotificationsIndicatorVisible(boolean visible) {
+        if (mNotificationsIndicator != null) {
+            mNotificationsIndicator.setVisibility(visible ? View.VISIBLE : View.GONE);
+            mNotificationsMenuItem.setIcon(visible
+                    ? R.drawable.icon_notifications_unread
+                    : R.drawable.icon_notifications);
+        }
     }
 
     @Override
@@ -145,6 +184,15 @@ public class HomeActivity extends BasePagerActivity implements
     @Override
     protected int getInitialLeftDrawerSelection(Menu menu) {
         mLeftDrawerMenu = menu;
+
+        mNotificationsMenuItem = menu.findItem(R.id.notifications);
+        if (mNotificationsMenuItem != null) {
+            View actionView = MenuItemCompat.getActionView(mNotificationsMenuItem);
+            mNotificationsIndicator =
+                    (ImageView) actionView.findViewById(R.id.notifications_indicator);
+            updateNotificationIndicator(mSelectedFactoryId);
+        }
+
         return mSelectedFactoryId;
     }
 
@@ -188,6 +236,8 @@ public class HomeActivity extends BasePagerActivity implements
     @Override
     public boolean onNavigationItemSelected(MenuItem item) {
         super.onNavigationItemSelected(item);
+
+        updateNotificationIndicator(item.getItemId());
 
         if (mFactory != null && mFactory.onDrawerItemSelected(item)) {
             return true;
@@ -334,6 +384,7 @@ public class HomeActivity extends BasePagerActivity implements
     @Override
     public void onRefresh() {
         forceLoaderReload(0);
+        forceLoaderReload(1);
         mFactory.onRefresh();
         super.onRefresh();
     }
